Update Fork to Master

+++ sky/backends/wheel_utils.py
@@ -17,14 +17,16 @@ import re
 import shutil
 import subprocess
 import tempfile
-from typing import Tuple
+from typing import Optional, Tuple
 
+import colorama
 import filelock
 from packaging import version
 
 import sky
 from sky import sky_logging
 from sky.backends import backend_utils
+from sky.server import common
 
 logger = sky_logging.init_logger(__name__)
 
@@ -39,30 +41,63 @@ _WHEEL_PATTERN = (f'{_PACKAGE_WHEEL_NAME}-'
                   f'{version.parse(sky.__version__)}-*.whl')
 
 
-def _get_latest_wheel_and_remove_all_others() -> pathlib.Path:
-    wheel_name = (f'**/{_WHEEL_PATTERN}')
+def _remove_stale_wheels(latest_wheel_dir: pathlib.Path) -> None:
+    """Remove all wheels except the latest one."""
+    for f in WHEEL_DIR.iterdir():
+        if f != latest_wheel_dir:
+            if f.is_dir() and not f.is_symlink():
+                shutil.rmtree(f, ignore_errors=True)
+
+
+def _get_latest_wheel() -> pathlib.Path:
+    wheel_name = f'**/{_WHEEL_PATTERN}'
     try:
         latest_wheel = max(WHEEL_DIR.glob(wheel_name), key=os.path.getctime)
     except ValueError:
         raise FileNotFoundError(
             'Could not find built SkyPilot wheels with glob pattern '
             f'{wheel_name} under {WHEEL_DIR!r}') from None
-
-    latest_wheel_dir_name = latest_wheel.parent
-    # Cleanup older wheels.
-    for f in WHEEL_DIR.iterdir():
-        if f != latest_wheel_dir_name:
-            if f.is_dir() and not f.is_symlink():
-                shutil.rmtree(f, ignore_errors=True)
     return latest_wheel
 
 
-def _build_sky_wheel():
-    """Build a wheel for SkyPilot."""
-    with tempfile.TemporaryDirectory() as tmp_dir:
+def _build_sky_wheel() -> pathlib.Path:
+    """Build a wheel for SkyPilot and return the path to the wheel."""
+    # Double check that the installed code is actually the same version as the
+    # running code. If not, the wheel we build will not match _WHEEL_PATTERN.
+    # See https://github.com/skypilot-org/skypilot/issues/5311.
+    version_on_disk = common.get_skypilot_version_on_disk()
+    if version_on_disk != sky.__version__:
+        logger.warning(
+            'Wheel build: The installed SkyPilot version is different from the '
+            'running code.\n'
+            f'{colorama.Style.DIM}'
+            f'running version: {sky.__version__}\n'
+            f'installed version: {version_on_disk}\n'
+            f'{colorama.Style.RESET_ALL}'
+            # The following message only applies to local API server. We have no
+            # way to tell from here if this is a remote or local API server. But
+            # we expect this to happen much more commonly to a local API server,
+            # so just print the hint regardless.
+            f'{colorama.Fore.YELLOW}'
+            'Please restart the local API server by running:\n'
+            f'{colorama.Style.BRIGHT}sky api stop; sky api start'
+            f'{colorama.Style.RESET_ALL}')
+        raise RuntimeError('The installed SkyPilot version is different from '
+                           'the running code. Please restart the SkyPilot API '
+                           'server with: sky api stop; sky api start')
+
+    with tempfile.TemporaryDirectory() as tmp_dir_str:
         # prepare files
-        tmp_dir = pathlib.Path(tmp_dir)
-        (tmp_dir / 'sky').symlink_to(SKY_PACKAGE_PATH, target_is_directory=True)
+        tmp_dir = pathlib.Path(tmp_dir_str)
+        sky_tmp_dir = tmp_dir / 'sky'
+        sky_tmp_dir.mkdir()
+        for item in SKY_PACKAGE_PATH.iterdir():
+            target = sky_tmp_dir / item.name
+            if item.name != '__init__.py':
+                # We do not symlink `sky/__init__.py` as we need to
+                # modify the commit hash in the file later.
+                # Symlink other files/folders.
+                target.symlink_to(item, target_is_directory=item.is_dir())
         setup_files_dir = SKY_PACKAGE_PATH / 'setup_files'
 
         setup_content = (setup_files_dir / 'setup.py').read_text()
@@ -76,6 +111,23 @@ def _build_sky_wheel():
         for f in setup_files_dir.iterdir():
             if f.is_file() and f.name != 'setup.py':
                 shutil.copy(str(f), str(tmp_dir))
+                if f.name == 'MANIFEST.in':
+                    # Remove the line `sky/dashboard/out`, so we do not
+                    # include the dashboard files in the internal wheel
+                    import fileinput  # pylint: disable=import-outside-toplevel
+                    with fileinput.input(tmp_dir / f.name,
+                                         inplace=True) as file:
+                        for line in file:
+                            if 'sky/dashboard/out' not in line:
+                                print(line, end='')
+
+        init_file_path = SKY_PACKAGE_PATH / '__init__.py'
+        init_file_content = init_file_path.read_text()
+        # Replace the commit hash with the current commit hash.
+        init_file_content = re.sub(
+            r'_SKYPILOT_COMMIT_SHA = [\'"](.*?)[\'"]',
+            f'_SKYPILOT_COMMIT_SHA = \'{sky.__commit__}\'', init_file_content)
+        (tmp_dir / 'sky' / '__init__.py').write_text(init_file_content)
 
         # It is important to normalize the path, otherwise 'pip wheel' would
         # treat the directory as a file and generate an empty wheel.
@@ -91,16 +143,16 @@ def _build_sky_wheel():
                            stderr=subprocess.PIPE,
                            check=True)
         except subprocess.CalledProcessError as e:
-            raise RuntimeError('Fail to build pip wheel for SkyPilot. '
+            raise RuntimeError('Failed to build pip wheel for SkyPilot. '
                                f'Error message: {e.stderr.decode()}') from e
 
         try:
             wheel_path = next(tmp_dir.glob(_WHEEL_PATTERN))
         except StopIteration:
             raise RuntimeError(
-                f'Fail to find pip wheel for SkyPilot under {tmp_dir} with '
+                f'Failed to find pip wheel for SkyPilot under {tmp_dir} with '
                 f'glob pattern {_WHEEL_PATTERN!r}. '
-                f'Found: {list(map(str, tmp_dir.glob("*")))}.'
+                f'Found: {list(map(str, tmp_dir.glob("*")))}. '
                 'No wheel file is generated.') from None
 
         # Use a unique temporary dir per wheel hash, because there may be many
@@ -112,7 +164,12 @@ def _build_sky_wheel():
 
         wheel_dir = WHEEL_DIR / hash_of_latest_wheel
         wheel_dir.mkdir(parents=True, exist_ok=True)
-        shutil.move(str(wheel_path), wheel_dir)
+        # shutil.move will fail when the file already exists and is being
+        # moved across filesystems.
+        if not os.path.exists(
+                os.path.join(wheel_dir, os.path.basename(wheel_path))):
+            shutil.move(str(wheel_path), wheel_dir)
+        return wheel_dir / wheel_path.name
 
 
 def build_sky_wheel() -> Tuple[pathlib.Path, str]:
@@ -127,13 +184,29 @@ def build_sky_wheel() -> Tuple[pathlib.Path, str]:
         - wheel_hash: The wheel content hash.
     """
 
-    def _get_latest_modification_time(path: pathlib.Path) -> float:
+    def _get_latest_modification_time(path: pathlib.Path) -> Optional[float]:
+        max_time = -1.
         if not path.exists():
-            return -1.
-        try:
-            return max(os.path.getmtime(root) for root, _, _ in os.walk(path))
-        except ValueError:
-            return -1.
+            return max_time
+        for root, dirs, files in os.walk(path):
+            # Prune __pycache__ directories to prevent walking into them and
+            # exclude them from processing
+            if '__pycache__' in dirs:
+                dirs.remove('__pycache__')
+            # Filter out .pyc files
+            filtered_files = [f for f in files if not f.endswith('.pyc')]
+            # Process remaining directories and files
+            for entry in (*dirs, *filtered_files):
+                entry_path = os.path.join(root, entry)
+                try:
+                    mtime = os.path.getmtime(entry_path)
+                    if mtime > max_time:
+                        max_time = mtime
+                except OSError:
+                    # Handle cases where file might have been deleted after
+                    # listing
+                    return None
+        return max_time
 
     # This lock prevents that the wheel is updated while being copied.
     # Although the current caller already uses a lock, we still lock it here
@@ -145,13 +218,24 @@ def build_sky_wheel() -> Tuple[pathlib.Path, str]:
         last_modification_time = _get_latest_modification_time(SKY_PACKAGE_PATH)
         last_wheel_modification_time = _get_latest_modification_time(WHEEL_DIR)
 
-        # only build wheels if the wheel is outdated
-        if last_wheel_modification_time < last_modification_time:
+        # Only build wheels if the wheel is outdated, wheel does not exist
+        # for the requested version, or files were deleted during checking.
+        if ((last_modification_time is None or
+             last_wheel_modification_time is None) or
+            (last_wheel_modification_time < last_modification_time) or
+                not any(WHEEL_DIR.glob(f'**/{_WHEEL_PATTERN}'))):
             if not WHEEL_DIR.exists():
                 WHEEL_DIR.mkdir(parents=True, exist_ok=True)
-            _build_sky_wheel()
+            latest_wheel = _build_sky_wheel()
+        else:
+            latest_wheel = _get_latest_wheel()
 
-        latest_wheel = _get_latest_wheel_and_remove_all_others()
+        # We remove all wheels except the latest one for garbage collection.
+        # Otherwise stale wheels will accumulate over time.
+        # TODO(romilb): If the user switches versions every alternate launch,
+        #  the wheel will be rebuilt every time. At the risk of adding
+        #  complexity, we can consider TTL caching wheels by version here.
+        _remove_stale_wheels(latest_wheel.parent)
 
         wheel_hash = latest_wheel.parent.name
